// Frontend/quran-dashboard-ui/src/app/core/api/generated/models/phrase-occurrence-dto.ts
/* eslint-disable */
/* Code generated by ng-openapi-gen DO NOT EDIT. */

import { PhraseAyahWordDto } from '../models/phrase-ayah-word-dto';
import { PhraseOccurrenceHighlightsDto } from '../models/phrase-occurrence-highlights-dto';
export interface PhraseOccurrenceDto {
  ayahId: number;
  ayahNumber: number;
  endWordNumber: number;
  highlights: PhraseOccurrenceHighlightsDto;
  occurrenceId: number;
  pageFrom: number;
  pageTo: number;
  startWordNumber: number;
  surahNameArabic: string;
  surahNumber: number;
  verseKey: string;
  words: Array<PhraseAyahWordDto>;
}